The tile ID limit has been expanded to 512.
Old worlds will be "locked" until i can figure out how to get them to safely convert.
Invalid blocks and items will be converted to air; this is necessary for the longevity of this project, plus future mod support.
Item IDs in old saves will be bumped up by 256 to prevet loss of items (you wont see this yet).
World options now save/load properly.
Fullbright is now a toggle.
The HUD coords display will now always be centered, regardless of aspect ratio.
Sneaking can now be toggled with [Caps Lock].
Dashing has been re-bound to [Tab].
The Host Options menu has been re-bound to [`/~].
Chat commands have been added.
The default block reach has been extended to 5 blocks (in Survival).
Fishing has been reworked: now has a loot table, similar to newer MC; the Aether has its own unique fishing loot table.
All projectiles now render correctly.
The Aether accessory slots and various Aether accessory items have been added.
Fossil Ores have been added to the Overworld, Nether, Aether, and Outer End.
Ore distribution has been improved in the Nether and Aether.
Aether worldgen has been greatly improved.
Nurealm worldgen has been greatly improved.
Nusa Tower loot has been improved.
The ladders in Nusa Towers have been fixed.
A new structure: abandoned tents; if you know, you know.
A new Overworld biome: the Red Desert; contains surfaces made of Stained Sand and Terracotta.
Tweaked Birch Forest and Lavender Forest generation.
Fixed an issue with the Outer End surface layer generation.
Added Blackroot: a new crop that grows on the underside of Netherrack; drops Blackroot Berries.
Added Magmaroot Berries: obtained by smelting Blackroot Berries.
Added Lava Reeds: a new crop that grows on Lava Rock beaches in the Nether; used to craft Nether Flax.
Nether Flax is used to craft String, and can be consumed (similar properties to Netherwart Bread).
Added Bone Vines: bone-like vines that generate throughout the Nether.
Netherrack spikes now generate throughout the Nether.
Aerclouds have been made fully accurate.
Gold Dungeons no longer get cut off by chunk borders.
Gold Dungeons now place a Treasure Chest instead of a regular Chest.
The Valkyrie toolset has been added, which increases your reach by +1 block.
Nethanium Swords and Axes now inflict fire.
All Stick recipes now give 8 Sticks instead of 4.
Endactuses are now used to craft Sticks.
Improved mob spawning rates in the Aether.
Added Moas, a tameable, rideable mob found in the Aether.
Added the Sun Spirit; a singular Sun Spirit spawns in the center of a Gold Dungeon, currently has no functionality, but it will be implemented soon enough.
Added Pigman trading; right-click a Pigman with a Gold Ingot to trade.
Improved the flight behavior of the Evupul, Dark Evupul, Nuskull, and Nuclear Nuskull.
Added tree-miner (use the Axe); all logs in a column (going upward from the destroyed log) will simultaneously be mined, damaging the Axe accordingly.
Many, many bugs from Hellish Ends 1.0 have been fixed.

+++ Minecraft.Client/SurvivalMode.cpp
@@ -64,7 +64,7 @@ bool SurvivalMode::destroyBlock(int x, int y, int z, int face)
     bool changed = GameMode::destroyBlock(x, y, z, face);
 
     shared_ptr<ItemInstance> item = minecraft->player->getSelectedItem();
-    bool couldDestroy = minecraft->player->canDestroy(Tile::tiles[t]);
+    bool couldDestroy = t > 0 && Tile::tiles[t] != nullptr && minecraft->player->canDestroy(Tile::tiles[t]);
     if (item != NULL)
 	{
         item->mineBlock(t, x, y, z, minecraft->player);
@@ -73,7 +73,7 @@ bool SurvivalMode::destroyBlock(int x, int y, int z, int face)
             minecraft->player->removeSelectedItem();
         }
     }
-    if (changed && couldDestroy) 
+    if (changed && couldDestroy && t > 0 && Tile::tiles[t] != nullptr)
 	{
 		Tile::tiles[t]->playerDestroy(minecraft->level, minecraft->player, x, y, z, data);
 	}
@@ -86,8 +86,8 @@ void SurvivalMode::startDestroyBlock(int x, int y, int z, int face)
 	if (!minecraft->player->mayBuild(x, y, z)) return;
     minecraft->level->extinguishFire(minecraft->player, x, y, z, face);
     int t = minecraft->level->getTile(x, y, z);
-    if (t > 0 && destroyProgress == 0) Tile::tiles[t]->attack(minecraft->level, x, y, z, minecraft->player);
-    if (t > 0 && Tile::tiles[t]->getDestroyProgress(minecraft->player) >= 1)
+    if (t > 0 && Tile::tiles[t] != nullptr && destroyProgress == 0) Tile::tiles[t]->attack(minecraft->level, x, y, z, minecraft->player);
+    if (t > 0 && Tile::tiles[t] != nullptr && Tile::tiles[t]->getDestroyProgress(minecraft->player) >= 1)
 	{
         destroyBlock(x, y, z, face);
     }
@@ -110,7 +110,7 @@ void SurvivalMode::continueDestroyBlock(int x, int y, int z, int face)
 	{
         int t = minecraft->level->getTile(x, y, z);
 		if (!minecraft->player->mayBuild(x, y, z)) return;
-        if (t == 0) return;
+        if (t == 0 || Tile::tiles[t] == nullptr) return;
         Tile *tile = Tile::tiles[t];
 
         destroyProgress += tile->getDestroyProgress(minecraft->player);
@@ -192,7 +192,7 @@ void SurvivalMode::tick()
 bool SurvivalMode::useItemOn(shared_ptr<Player> player, Level *level, shared_ptr<ItemInstance> item, int x, int y, int z, int face, bool bTestUseOnOnly, bool *pbUsedItem)
 {
 	int t = level->getTile(x, y, z);
-	if (t > 0)
+	if (t > 0 && Tile::tiles[t] != nullptr)
 	{
 		if (Tile::tiles[t]->use(level, x, y, z, player)) return true;
 	}
